reorged TokensPage
