+++ site/src/pages/UserSettingsPage/TokensPage/components/index.ts
@@ -1,0 +1,2 @@
+export { ConfirmDeleteDialog } from "./ConfirmDeleteDialog"
+export { TokensSwitch } from "./TokensSwitch"
